Data.Conduit.Resumable

Resumable sources

Resumable sources are provided by the conduit package itself. See ResumableSource.

newResumableSource :: Monad m => Source m o -> ResumableSource m o

Convert a Source into a ResumableSource so it can be used with $$++.

Resumable sinks

There is no "ResumableSink" type because Sink does not require finalizer support. A Sink is finalized by letting it run to completion with $$.

(+$$) :: Monad m => Source m i -> Sink i m r -> m (Either (Sink i m r) r)

Connect a source and a sink, allowing the sink to be fed more data later. Return a Right if the sink completes, or a Left if the source is exhausted and the sink requests more input.

When you are done with the sink, close it with $$ so that:

The sink sees the end of stream. The sink never sees await return Nothing until you finish the sink with $$.

The sink can release system resources.

Combining with resumable sources

(+$$+) :: Monad m => Source m i -> Sink i m r -> m (ResumableSource m i, Either (Sink i m r) r)

Connect a source to a sink, allowing both to be resumed.

(+$$++) :: Monad m => ResumableSource m i -> Sink i m r -> m (ResumableSource m i, Either (Sink i m r) r)

Like +$$+, but resume an already-running source.

(+$$+-) :: Monad m => ResumableSource m i -> Sink i m r -> m (Either (Sink i m r) r)

Finish processing a ResumableSource, but allow the Sink to be reused.

Resumable conduits

data ResumableConduit i m o

newResumableConduit :: Monad m => Conduit i m o -> ResumableConduit i m o

Convert a Conduit into a ResumableConduit so it can be used with =$++.

(=$+) :: Monad m => Conduit a m b -> Sink b m r -> Sink a m (ResumableConduit a m b, r)

Fuse a conduit behind a sink, but allow the conduit to be reused after the sink returns.

When the source runs out, the stream terminator is sent directly to the sink, bypassing the conduit. Some conduits wait for a stream terminator before producing their remaining output, so be sure to use =$+- to "flush" this data out.

(=$++) :: Monad m => ResumableConduit a m b -> Sink b m r -> Sink a m (ResumableConduit a m b, r)

Continue using a conduit after =$+.

(=$+-) :: Monad m => ResumableConduit a m b -> Sink b m r -> Sink a m r

Finalize a ResumableConduit by using it one more time. It will be closed when the sink finishes.